+++ src/components/gridContainer/gridContainer.component.ts
@@ -1,7 +1,6 @@
-import {Component, ChangeDetectionStrategy, ViewContainerRef, ViewChild, inject, OnInit} from '@angular/core';
+import {Component, ChangeDetectionStrategy, OnInit} from '@angular/core';
 
-import {ContentRendererCurrentViewContainer} from '../../interfaces';
-import {CONTENT_RENDERER_CURRENT_VIEW_CONTAINER} from '../../misc/tokens';
+import {RenderableContentComponent} from '../renderableContent/renderableContent.component';
 
 /**
  * Component that represents grid container
@@ -9,27 +8,12 @@ import {CONTENT_RENDERER_CURRENT_VIEW_CONTAINER} from '../../misc/tokens';
 @Component(
 {
     selector: '[gridContainer]',
-    templateUrl: 'gridContainer.component.html',
+    templateUrl: '../renderableContent/renderableContent.component.html',
     standalone: true,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class GridContainerSAComponent implements OnInit
+export class GridContainerSAComponent extends RenderableContentComponent implements OnInit
 {
-    //######################### protected fields #########################
-
-    /**
-     * Instance of class that is used for sharing current view container
-     */
-    protected contentRendererViewContainer: ContentRendererCurrentViewContainer = inject(CONTENT_RENDERER_CURRENT_VIEW_CONTAINER);
-
-    //######################### protected properties - children #########################
-
-    /**
-     * Instance of view container for content container content
-     */
-    @ViewChild('container', {static: true, read: ViewContainerRef})
-    protected contentViewContainer!: ViewContainerRef;
-
     //######################### public methods - implementation of OnInit #########################
     
     /**
@@ -37,6 +21,6 @@ export class GridContainerSAComponent implements OnInit
      */
     public ngOnInit(): void
     {
-        this.contentRendererViewContainer.viewContainer = this.contentViewContainer;
+        this.contentRendererInnerStructure.gridContainer.renderableContent = this;
     }
 }