+++ src/components/index.ts
@@ -2,5 +2,6 @@ export * from './contentContainer/contentContainer.component';
 export * from './grid/grid.component';
 export * from './gridContainer/gridContainer.component';
 export * from './matrixGrid/matrixGrid.component';
+export * from './renderableContent/renderableContent.component';
 export * from './tableGridColumn/tableGridColumn.component';
 export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.component';