+++ src/plugins/contentRenderer/matrix/misc/components/index.ts
@@ -1,0 +1,3 @@
+export * from './baseDefaultTemplates/baseDefaultTemplates.component';
+export * from './cssGridDefaultTemplates/cssGridDefaultTemplates.component';
+export * from './tableDefaultTemplates/tableDefaultTemplates.component';